zynqmp: Allow 32-bit free-running clock
The current zynq machine code assumes a 16-bit counter, which is only
true for the zynq7000. The zynqmp has a 32-bit counter. The timestamp
timer would stop functioning on the zynqmp after it overflowed. This
commit handles the overflow properly on the zynqmp.

+++ libplatsupport/mach_include/zynq/platsupport/mach/timer.h
@@ -174,4 +174,4 @@ void ttc_freerun(ttc_t *tcc);
 int ttc_handle_irq(ttc_t *ttc);
 uint64_t ttc_get_time(ttc_t *ttc);
 /* convert from a ticks value to ns for a configured ttc */
-uint64_t ttc_ticks_to_ns(ttc_t *ttc, uint16_t ticks);
+uint64_t ttc_ticks_to_ns(ttc_t *ttc, uint32_t ticks);
